+++ src/main/java/io/github/talelin/latticy/controller/cms/AdminController.java
@@ -6,16 +6,34 @@ import io.github.talelin.core.annotation.PermissionMeta;
 import io.github.talelin.core.annotation.PermissionModule;
 import io.github.talelin.latticy.bo.GroupPermissionBO;
 import io.github.talelin.latticy.common.util.PageUtil;
-import io.github.talelin.latticy.dto.admin.*;
+import io.github.talelin.latticy.dto.admin.DispatchPermissionDTO;
+import io.github.talelin.latticy.dto.admin.DispatchPermissionsDTO;
+import io.github.talelin.latticy.dto.admin.NewGroupDTO;
+import io.github.talelin.latticy.dto.admin.RemovePermissionsDTO;
+import io.github.talelin.latticy.dto.admin.ResetPasswordDTO;
+import io.github.talelin.latticy.dto.admin.UpdateGroupDTO;
+import io.github.talelin.latticy.dto.admin.UpdateUserInfoDTO;
 import io.github.talelin.latticy.model.GroupDO;
 import io.github.talelin.latticy.model.PermissionDO;
 import io.github.talelin.latticy.model.UserDO;
 import io.github.talelin.latticy.service.AdminService;
 import io.github.talelin.latticy.service.GroupService;
-import io.github.talelin.latticy.vo.*;
+import io.github.talelin.latticy.vo.CreatedVO;
+import io.github.talelin.latticy.vo.DeletedVO;
+import io.github.talelin.latticy.vo.PageResponseVO;
+import io.github.talelin.latticy.vo.UpdatedVO;
+import io.github.talelin.latticy.vo.UserInfoVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
@@ -26,6 +44,7 @@ import java.util.stream.Collectors;
 
 /**
  * @author pedro@TaleLin
+ * @author Juzi@TaleLin
  */
 @RestController
 @RequestMapping("/cms/admin")
@@ -52,12 +71,12 @@ public class AdminController {
     @PermissionMeta(value = "查询所有用户", mount = false)
     public PageResponseVO getUsers(
             @RequestParam(name = "group_id", required = false)
-            @Min(value = 1, message = "{group.id.positive}") Long groupId,
+            @Min(value = 1, message = "{group.id.positive}") Integer groupId,
             @RequestParam(name = "count", required = false, defaultValue = "10")
             @Min(value = 1, message = "{page.count.min}")
-            @Max(value = 30, message = "{page.count.max}") Long count,
+            @Max(value = 30, message = "{page.count.max}") Integer count,
             @RequestParam(name = "page", required = false, defaultValue = "0")
-            @Min(value = 0, message = "{page.number.min}") Long page) {
+            @Min(value = 0, message = "{page.number.min}") Integer page) {
         IPage<UserDO> iPage = adminService.getUserPageByGroupId(groupId, count, page);
         List<UserInfoVO> userInfos = iPage.getRecords().stream().map(user -> {
             List<GroupDO> groups = groupService.getUserGroupsByUserId(user.getId());
@@ -69,7 +88,7 @@ public class AdminController {
     @PutMapping("/user/{id}/password")
     @AdminRequired
     @PermissionMeta(value = "修改用户密码", mount = false)
-    public UpdatedVO changeUserPassword(@PathVariable @Positive(message = "{id.positive}") Long id, @RequestBody @Validated ResetPasswordDTO validator) {
+    public UpdatedVO changeUserPassword(@PathVariable @Positive(message = "{id.positive}") Integer id, @RequestBody @Validated ResetPasswordDTO validator) {
         adminService.changeUserPassword(id, validator);
         return new UpdatedVO(4);
     }
@@ -77,7 +96,7 @@ public class AdminController {
     @DeleteMapping("/user/{id}")
     @AdminRequired
     @PermissionMeta(value = "删除用户", mount = false)
-    public DeletedVO deleteUser(@PathVariable @Positive(message = "{id.positive}") Long id) {
+    public DeletedVO deleteUser(@PathVariable @Positive(message = "{id.positive}") Integer id) {
         adminService.deleteUser(id);
         return new DeletedVO(5);
     }
@@ -85,7 +104,7 @@ public class AdminController {
     @PutMapping("/user/{id}")
     @AdminRequired
     @PermissionMeta(value = "管理员更新用户信息", mount = false)
-    public UpdatedVO updateUser(@PathVariable @Positive(message = "{id.positive}") Long id, @RequestBody @Validated UpdateUserInfoDTO validator) {
+    public UpdatedVO updateUser(@PathVariable @Positive(message = "{id.positive}") Integer id, @RequestBody @Validated UpdateUserInfoDTO validator) {
         adminService.updateUserInfo(id, validator);
         return new UpdatedVO(6);
     }
@@ -96,9 +115,9 @@ public class AdminController {
     public PageResponseVO getGroups(
             @RequestParam(name = "count", required = false, defaultValue = "10")
             @Min(value = 1, message = "{page.count.min}")
-            @Max(value = 30, message = "{page.count.max}") Long count,
+            @Max(value = 30, message = "{page.count.max}") Integer count,
             @RequestParam(name = "page", required = false, defaultValue = "0")
-            @Min(value = 0, message = "{page.number.min}") Long page) {
+            @Min(value = 0, message = "{page.number.min}") Integer page) {
         IPage<GroupDO> iPage = adminService.getGroupPage(page, count);
         return PageUtil.build(iPage);
     }
@@ -114,7 +133,7 @@ public class AdminController {
     @GetMapping("/group/{id}")
     @AdminRequired
     @PermissionMeta(value = "查询一个权限组及其权限", mount = false)
-    public GroupPermissionBO getGroup(@PathVariable @Positive(message = "{id.positive}") Long id) {
+    public GroupPermissionBO getGroup(@PathVariable @Positive(message = "{id.positive}") Integer id) {
         GroupPermissionBO groupPermissions = adminService.getGroup(id);
         return groupPermissions;
     }
@@ -130,7 +149,7 @@ public class AdminController {
     @PutMapping("/group/{id}")
     @AdminRequired
     @PermissionMeta(value = "更新一个权限组", mount = false)
-    public UpdatedVO updateGroup(@PathVariable @Positive(message = "{id.positive}") Long id,
+    public UpdatedVO updateGroup(@PathVariable @Positive(message = "{id.positive}") Integer id,
                                        @RequestBody @Validated UpdateGroupDTO validator) {
         adminService.updateGroup(id, validator);
         return new UpdatedVO(7);
@@ -139,7 +158,7 @@ public class AdminController {
     @DeleteMapping("/group/{id}")
     @AdminRequired
     @PermissionMeta(value = "删除一个权限组", mount = false)
-    public DeletedVO deleteGroup(@PathVariable @Positive(message = "{id.positive}") Long id) {
+    public DeletedVO deleteGroup(@PathVariable @Positive(message = "{id.positive}") Integer id) {
         adminService.deleteGroup(id);
         return new DeletedVO(8);
     }
