+++ src/main/java/io/github/talelin/latticy/service/impl/LogServiceImpl.java
@@ -1,44 +1,45 @@
 package io.github.talelin.latticy.service.impl;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
-import io.github.talelin.latticy.common.mybatis.Page;
-import io.github.talelin.latticy.model.LogDO;
-import io.github.talelin.latticy.mapper.LogMapper;
-import io.github.talelin.latticy.service.LogService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import io.github.talelin.latticy.common.mybatis.Page;
+import io.github.talelin.latticy.mapper.LogMapper;
+import io.github.talelin.latticy.model.LogDO;
+import io.github.talelin.latticy.service.LogService;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
 
 /**
  * @author pedro@TaleLin
+ * @author Juzi@TaleLin
  */
 @Service
 public class LogServiceImpl extends ServiceImpl<LogMapper, LogDO> implements LogService {
 
     @Override
-    public IPage<LogDO> getLogPage(Long page, Long count, String name, Date start, Date end) {
+    public IPage<LogDO> getLogPage(Integer page, Integer count, String name, Date start, Date end) {
         Page<LogDO> pager = new Page<>(page, count);
         IPage<LogDO> iPage = this.baseMapper.findLogsByUsernameAndRange(pager, name, start, end);
         return iPage;
     }
 
     @Override
-    public IPage<LogDO> searchLogPage(Long page, Long count, String name, String keyword, Date start, Date end) {
+    public IPage<LogDO> searchLogPage(Integer page, Integer count, String name, String keyword, Date start, Date end) {
         Page<LogDO> pager = new Page<>(page, count);
         IPage<LogDO> iPage = this.baseMapper.searchLogsByUsernameAndKeywordAndRange(pager, name, "%" + keyword + "%", start, end);
         return iPage;
     }
 
     @Override
-    public IPage<String> getUserNamePage(Long page, Long count) {
+    public IPage<String> getUserNamePage(Integer page, Integer count) {
         Page<LogDO> pager = new Page<>(page, count);
         IPage<String> iPage = this.baseMapper.getUserNames(pager);
         return iPage;
     }
 
     @Override
-    public boolean createLog(String message, String permission, Long userId, String username, String method, String path, Integer status) {
+    public boolean createLog(String message, String permission, Integer userId, String username, String method, String path, Integer status) {
         LogDO record = LogDO.builder()
                 .message(message)
                 .userId(userId)
